Refactor mean-field guides.
